feat: :art: Separation of map concerns
Moved map listeners to map script file
